+++ app/src/main/java/com/music/player/bhandari/m/model/dataItem.java
@@ -21,7 +21,7 @@ import androidx.annotation.NonNull;
 public class dataItem {
 
 
-    public dataItem(int id, String title, int artist_id, String artist_name, long album_id, String albumName
+    public dataItem(int id, String title, long artist_id, String artist_name, long album_id, String albumName
             , String year, String file_path, String duration, int trackNumber) {
         if (title != null) {
             this.title = title;
@@ -55,7 +55,7 @@ public class dataItem {
         this.trackNumber = trackNumber;
     }
 
-    public dataItem(int id, String title, int numberOfTracks, int numberOfAlbums) {
+    public dataItem(long id, String title, int numberOfTracks, int numberOfAlbums) {
 
         this.artist_id = id;
         if (title != null) {
@@ -68,7 +68,7 @@ public class dataItem {
     }
 
 
-    public dataItem(long id, String title, String artist_name, int numberOfTracks, String year, int artist_id) {
+    public dataItem(long id, String title, String artist_name, int numberOfTracks, String year, long artist_id) {
         this.artist_name = artist_name;
         this.artist_id = artist_id;
         if (title != null) {
@@ -97,7 +97,7 @@ public class dataItem {
 
     public String title = "";
 
-    public int artist_id;
+    public long artist_id;
     public String artist_name = "";
 
     public long album_id;
